code names and such are now 100% consistent with UML diagram

+++ src/main/java/csulb/cecs323/model/Food.java
@@ -5,7 +5,7 @@ import javax.persistence.*;
 @Entity
 public class Food {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long foodId;
+    private int foodId;
     @Column(nullable=false)
     private String name;
     @Column(nullable=false)
